requirements-dev.txt
setuptools
setuptools-rust
wheel
Pillow
pytest
piexif
twine
check-manifest
